+++ packages/amplify-codegen-e2e-tests/src/__tests__/push-codegen-admin-modelgen.test.ts
@@ -1,0 +1,19 @@
+import { DEFAULT_JS_CONFIG, createNewProjectDir } from "@aws-amplify/amplify-codegen-e2e-core";
+import { deleteAmplifyProject, testPushAdminModelgen, testPushCodegen } from "../codegen-tests-base";
+
+const schema = 'admin-modelgen.graphql';
+
+describe('Amplify push with codegen tests - admin modelgen', () => {
+  let projectRoot: string;
+  beforeEach(async () => {
+    projectRoot = await createNewProjectDir('pushCodegenAdminModelgen');
+  });
+
+  afterEach(async () => {
+    await deleteAmplifyProject(projectRoot);
+  });
+
+  it(`should not throw error for executing the admin modelgen step required by studio CMS usage post push given the schema with input, union and interface types`, async () => {
+    await testPushAdminModelgen(DEFAULT_JS_CONFIG, projectRoot, schema);
+  });
+});