+++ packages/appsync-modelgen-plugin/src/visitors/appsync-model-introspection-visitor.ts
@@ -1,12 +1,16 @@
 import { DEFAULT_SCALARS, NormalizedScalarsMap } from "@graphql-codegen/visitor-plugin-common";
 import { GraphQLSchema } from "graphql";
-import { Argument, AssociationType, Field, Fields, FieldType, ModelAttribute, ModelIntrospectionSchema, PrimaryKeyInfo, SchemaEnum, SchemaModel, SchemaMutation, SchemaNonModel, SchemaQuery, SchemaSubscription } from "../interfaces/introspection";
+import { Argument, AssociationType, Field, Fields, FieldType, ModelAttribute, ModelIntrospectionSchema, PrimaryKeyInfo, SchemaEnum, SchemaModel, SchemaMutation, SchemaNonModel, SchemaQuery, SchemaSubscription, Input, InputFieldType } from "../interfaces/introspection";
 import { METADATA_SCALAR_MAP } from "../scalars";
 import { CodeGenConnectionType } from "../utils/process-connections";
-import { RawAppSyncModelConfig, ParsedAppSyncModelConfig, AppSyncModelVisitor, CodeGenEnum, CodeGenField, CodeGenModel, CodeGenPrimaryKeyType, CodeGenQuery, CodeGenSubscription, CodeGenMutation } from "./appsync-visitor";
+import { RawAppSyncModelConfig, ParsedAppSyncModelConfig, AppSyncModelVisitor, CodeGenEnum, CodeGenField, CodeGenModel, CodeGenPrimaryKeyType, CodeGenQuery, CodeGenSubscription, CodeGenMutation, CodeGenInputObject, CodeGenUnion, CodeGenInterface } from "./appsync-visitor";
+import fs from 'fs';
 import path from 'path';
 import Ajv from 'ajv';
 import modelIntrospectionSchema from '../schemas/introspection/1/ModelIntrospectionSchema.json';
+
+type UnionFieldType = { union: string };
+type InterfaceFieldType = { interface: string };
 
 export interface RawAppSyncModelIntrospectionConfig extends RawAppSyncModelConfig {};
 export interface ParsedAppSyncModelIntrospectionConfig extends ParsedAppSyncModelConfig {};
@@ -63,22 +67,46 @@ export class AppSyncModelIntrospectionVisitor<
     }, {});
     result = { ...result, models, nonModels, enums };
     const queries = Object.values(this.queryMap).reduce((acc, queryObj: CodeGenQuery) => {
+      // Skip the field if the field type is union/interface
+      // TODO: Remove this skip once these types are supported for stakeholder usages
+      const fieldType = this.getType(queryObj.type) as any;
+      if (this.isUnionFieldType(fieldType) || this.isInterfaceFieldType(fieldType)) {
+        return acc;
+      }
       return { ...acc, [queryObj.name]: this.generateGraphQLOperationMetadata<CodeGenQuery, SchemaQuery>(queryObj) };
     }, {})
     const mutations = Object.values(this.mutationMap).reduce((acc, mutationObj: CodeGenMutation) => {
+      // Skip the field if the field type is union/interface
+      // TODO: Remove this skip once these types are supported for stakeholder usages
+      const fieldType = this.getType(mutationObj.type) as any;
+      if (this.isUnionFieldType(fieldType) || this.isInterfaceFieldType(fieldType)) {
+        return acc;
+      }
       return { ...acc, [mutationObj.name]: this.generateGraphQLOperationMetadata<CodeGenMutation, SchemaMutation>(mutationObj) };
     }, {});
     const subscriptions = Object.values(this.subscriptionMap).reduce((acc, subscriptionObj: CodeGenSubscription) => {
+      // Skip the field if the field type is union/interface
+      // TODO: Remove this skip once these types are supported for stakeholder usages
+      const fieldType = this.getType(subscriptionObj.type) as any;
+      if (this.isUnionFieldType(fieldType) || this.isInterfaceFieldType(fieldType)) {
+        return acc;
+      }
       return { ...acc, [subscriptionObj.name]: this.generateGraphQLOperationMetadata<CodeGenSubscription, SchemaSubscription>(subscriptionObj) };
-    }, {})
-    if(Object.keys(queries).length > 0) {
+    }, {});
+    const inputs = Object.values(this.inputObjectMap).reduce((acc, inputObj: CodeGenInputObject) => {
+      return { ...acc, [inputObj.name]: this.generateGraphQLInputMetadata(inputObj) };
+    }, {});
+    if (Object.keys(queries).length > 0) {
       result = { ...result, queries };
     }
-    if(Object.keys(mutations).length > 0) {
+    if (Object.keys(mutations).length > 0) {
       result = { ...result, mutations };
     }
-    if(Object.keys(subscriptions).length > 0) {
+    if (Object.keys(subscriptions).length > 0) {
       result = { ...result, subscriptions };
+    }
+    if (Object.keys(inputs).length > 0) {
+      result = { ...result, inputs }
     }
     return result;
   }
@@ -119,10 +147,16 @@ export class AppSyncModelIntrospectionVisitor<
     return {
       name: this.getModelName(nonModel),
       fields: nonModel.fields.reduce((acc: Fields, field: CodeGenField) => {
+        // Skip the field if the field type is union/interface
+        // TODO: Remove this skip once these types are supported for stakeholder usages
+        const fieldType = this.getType(field.type) as any;
+        if (this.isUnionFieldType(fieldType) || this.isInterfaceFieldType(fieldType)) {
+          return acc;
+        }
         const fieldMeta: Field = {
           name: this.getFieldName(field),
           isArray: field.isList,
-          type: this.getType(field.type),
+          type: fieldType,
           isRequired: !field.isNullable,
           attributes: [],
         };
@@ -150,6 +184,30 @@ export class AppSyncModelIntrospectionVisitor<
       values: Object.values(enumObj.values),
     };
   }
+
+  /**
+   * Generate GraqhQL input object type metadata in model introspection schema from the codegen MIPR
+   * @param inputObj input type object
+   * @returns input type object metadata in model introspection schema
+   */
+  private generateGraphQLInputMetadata(inputObj: CodeGenInputObject): Input {
+    return {
+      name: inputObj.name,
+      attributes: inputObj.inputValues.reduce((acc, param ) => {
+        const arg: Argument = {
+          name: param.name,
+          isArray: param.isList,
+          type: this.getType(param.type) as InputFieldType,
+          isRequired: !param.isNullable
+        };
+        if (param.isListNullable !== undefined) {
+          arg.isArrayNullable = param.isListNullable;
+        }
+        return { ...acc, [param.name]: arg };
+      }, {}),
+    }
+  }
+
   /**
    * Generate GraqhQL operation (query/mutation/subscription) metadata in model introspection schema from the codegen MIPR
    * @param operationObj operation object
@@ -170,7 +228,7 @@ export class AppSyncModelIntrospectionVisitor<
         const arg: Argument = {
           name: param.name,
           isArray: param.isList,
-          type: this.getType(param.type),
+          type: this.getType(param.type) as InputFieldType,
           isRequired: !param.isNullable
         };
         if (param.isListNullable !== undefined) {
@@ -182,7 +240,7 @@ export class AppSyncModelIntrospectionVisitor<
     return operationMeta as V;
   }
 
-  protected getType(gqlType: string): FieldType {
+  protected getType(gqlType: string): FieldType | InputFieldType | UnionFieldType | InterfaceFieldType {
     // Todo: Handle unlisted scalars
     if (gqlType in METADATA_SCALAR_MAP) {
       return METADATA_SCALAR_MAP[gqlType] as FieldType;
@@ -196,7 +254,16 @@ export class AppSyncModelIntrospectionVisitor<
     if (gqlType in this.modelMap) {
       return { model: gqlType };
     }
-    throw new Error(`Unknown type ${gqlType}`);
+    if (gqlType in this.inputObjectMap) {
+      return { input: gqlType }
+    }
+    if (gqlType in this.unionMap) {
+      return { union: gqlType }
+    }
+    if (gqlType in this.interfaceMap) {
+      return { interface: gqlType }
+    }
+    throw new Error(`Unknown type ${gqlType} found during model introspection schema generation`);
   }
 
   private generateModelPrimaryKeyInfo(model: CodeGenModel): PrimaryKeyInfo {
@@ -210,5 +277,12 @@ export class AppSyncModelIntrospectionVisitor<
       };
     }
     throw new Error(`No primary key found for model ${model.name}`);
+  }
+
+  private isUnionFieldType = (obj: any): obj is UnionFieldType => {
+    return typeof obj === 'object' && typeof obj.union === 'string';
+  }
+  private isInterfaceFieldType = (obj: any): obj is InterfaceFieldType => {
+    return typeof obj === 'object' && typeof obj.interface === 'string';
   }
 }