+++ packages/graphql-generator/src/typescript.ts
@@ -14,12 +14,13 @@ export type GenerateTypesOptions = {
   queries: string | Source[];
   introspection?: boolean;
   multipleSwiftFiles?: boolean; // only used when target is swift
+  amplifyJsLibraryVersion?: number; // only used when target is angular
 };
 
 export type GenerateModelsOptions = {
   schema: string;
   target: ModelsTarget;
-  directives: string;
+  directives?: string;
   isDataStoreEnabled?: boolean;
 
   // feature flags